requirements.txt

Flask
requests
Pillow
